refactor[#523] : MyInfo types 폴더로 이동
