+++ src/entities/auth/api/getMyInfo.ts
@@ -1,18 +1,8 @@
 import { useQuery } from "@tanstack/react-query";
-import type { Region } from "@/entities/map/@x/auth";
 import { apiClient } from "@/shared/lib";
-import { MY_INFO_END_POINT, SOCIAL_TYPE } from "../constants";
+import { MY_INFO_END_POINT } from "../constants";
 import { authQueryKey } from "../constants";
-
-export interface MyInfo {
-  email: string;
-  gender: "FEMALE" | "MALE" | "NONE";
-  age: "10" | "20" | "30" | "40" | "50" | "60";
-  regions: Region[];
-  nickLastModDt: string | null;
-  socialType: keyof typeof SOCIAL_TYPE;
-  isPasswordSet: boolean;
-}
+import type { MyInfo } from "../types/server";
 
 type GetMyInfoResponse = MyInfo;
