+++ src/features/setting/api/putChangeGender.ts
@@ -1,7 +1,7 @@
 import { useEffect } from "react";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import type { MyInfo } from "@/entities/auth/api";
 import { authQueryKey } from "@/entities/auth/constants";
+import type { MyInfo } from "@/entities/auth/types/server";
 import { apiClient } from "@/shared/lib";
 import { SETTING_END_POINT } from "../constants";
