+++ src/pages/setting/manage-account/page.tsx
@@ -2,9 +2,9 @@ import { useQueryClient } from "@tanstack/react-query";
 import { AccountCancellationModal } from "@/features/setting/ui";
 import { PasswordChangeModal } from "@/features/setting/ui";
 import { PasswordSetModal } from "@/features/setting/ui";
-import type { MyInfo } from "@/entities/auth/api";
 import { useGetMyInfo } from "@/entities/auth/api";
 import { SOCIAL_TYPE } from "@/entities/auth/constants";
+import type { MyInfo } from "@/entities/auth/types/server";
 import { useModal, withAuth } from "@/shared/lib";
 import { DividerLine } from "@/shared/ui/divider";
 import { ArrowRightIcon } from "@/shared/ui/icon";
